refactor : Piece 움직임 추상화 수준 변경
Piece 가 바로 Movable interface 를 구현해 canMove 메소드를 오버라이딩하도록 수정
