+++ src/main/java/chess/domain/piece/Bishop.java
@@ -1,11 +1,11 @@
 package chess.domain.piece;
 
 import chess.domain.camp.TeamColor;
-import chess.domain.move.BishopMove;
+
+import java.util.Set;
 
 public class Bishop extends Piece {
-
-    private final BishopMove bishopMove = new BishopMove();
+    private static final int BISHOP_MAX_MOVE_COUNT = 8;
 
     public Bishop(final PieceType pieceType, final TeamColor teamColor) {
         super(pieceType, teamColor);
@@ -13,6 +13,7 @@ public class Bishop extends Piece {
 
     @Override
     public boolean canMove(final Position source, final Position target) {
-        return bishopMove.canMove(source, target);
+        final Set<Position> allPositions = move.getAllPositions(source, Direction.getDiagonalDirections(), BISHOP_MAX_MOVE_COUNT);
+        return allPositions.contains(target);
     }
 }
