+++ src/main/java/chess/domain/piece/King.java
@@ -1,11 +1,11 @@
 package chess.domain.piece;
 
 import chess.domain.camp.TeamColor;
-import chess.domain.move.KingMove;
+
+import java.util.Set;
 
 public class King extends Piece {
-
-    private static final KingMove kingMove = new KingMove();
+    private static final int KING_MAX_MOVE_COUNT = 1;
 
     public King(final PieceType pieceType, final TeamColor teamColor) {
         super(pieceType, teamColor);
@@ -13,6 +13,7 @@ public class King extends Piece {
 
     @Override
     public boolean canMove(final Position source, final Position target) {
-        return kingMove.canMove(source, target);
+        final Set<Position> allPositions = move.getAllPositions(source, Direction.getAllDirections(), KING_MAX_MOVE_COUNT);
+        return allPositions.contains(target);
     }
 }
