+++ src/main/java/chess/domain/piece/Movable.java
@@ -1,8 +1,8 @@
-package chess.domain.move;
-
-import chess.domain.piece.Position;
+package chess.domain.piece;
 
 @FunctionalInterface
 public interface Movable {
+    Move move = new Move();
+
     boolean canMove(final Position source, final Position target);
 }
