+++ src/main/java/chess/domain/piece/Queen.java
@@ -1,11 +1,11 @@
 package chess.domain.piece;
 
 import chess.domain.camp.TeamColor;
-import chess.domain.move.QueenMove;
+
+import java.util.Set;
 
 public class Queen extends Piece {
-
-    private final QueenMove queenMove = new QueenMove();
+    private static final int QUEEN_MAX_MOVE_COUNT = 8;
 
     public Queen(final PieceType pieceType, final TeamColor teamColor) {
         super(pieceType, teamColor);
@@ -13,6 +13,7 @@ public class Queen extends Piece {
 
     @Override
     public boolean canMove(final Position source, final Position target) {
-        return queenMove.canMove(source, target);
+        final Set<Position> allPositions = move.getAllPositions(source, Direction.getAllDirections(), QUEEN_MAX_MOVE_COUNT);
+        return allPositions.contains(target);
     }
 }
