+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,11 +1,11 @@
 package chess.domain.piece;
 
 import chess.domain.camp.TeamColor;
-import chess.domain.move.RookMove;
+
+import java.util.Set;
 
 public class Rook extends Piece {
-
-    private final RookMove rookMove = new RookMove();
+    private static final int ROOK_MAX_MOVE_COUNT = 8;
 
     public Rook(final PieceType pieceType, final TeamColor teamColor) {
         super(pieceType, teamColor);
@@ -13,6 +13,7 @@ public class Rook extends Piece {
 
     @Override
     public boolean canMove(final Position source, final Position target) {
-        return rookMove.canMove(source, target);
+        final Set<Position> allPositions = move.getAllPositions(source, Direction.getFourDirections(), ROOK_MAX_MOVE_COUNT);
+        return allPositions.contains(target);
     }
 }
